Move swap types to separate file
