+++ src/components/swap/SwapButtons/SwapButtons.tsx
@@ -15,7 +15,7 @@ import { useActiveWeb3React } from '../../../hooks'
 import { ApprovalState } from '../../../hooks/useApproveCallback'
 import { WrapState, WrapType } from '../../../hooks/useWrapCallback'
 import { SwapData } from '../../../pages/Swap'
-import { Field } from '../../../state/swap/actions'
+import { Field } from '../../../state/swap/types'
 import { useIsExpertMode } from '../../../state/user/hooks'
 import { warningSeverity } from '../../../utils/prices'
 import { ButtonConfirmed, ButtonError, ButtonPrimary } from '../../Button'
