+++ src/components/swap/SwapPlatformSelector.tsx
@@ -12,8 +12,8 @@ import useDebounce from '../../hooks/useDebounce'
 import { useGasFeesUSD } from '../../hooks/useGasFeesUSD'
 import { useIsMobileByMedia } from '../../hooks/useIsMobileByMedia'
 import { useSwapsGasEstimations } from '../../hooks/useSwapsGasEstimate'
-import { Field } from '../../state/swap/actions'
 import { useSwapState } from '../../state/swap/hooks'
+import { Field } from '../../state/swap/types'
 import { useUserSlippageTolerance } from '../../state/user/hooks'
 import { TYPE } from '../../theme'
 import {
