+++ src/hooks/useSwapsGasEstimate.ts
@@ -8,8 +8,8 @@ import { INITIAL_ALLOWED_SLIPPAGE } from '../constants'
 import { useTokenAllowancesForMultipleSpenders } from '../data/Allowances'
 import { MainnetGasPrice } from '../state/application/actions'
 import { useMainnetGasPrices } from '../state/application/hooks'
-import { Field } from '../state/swap/actions'
 import { tryParseAmount, useSwapState } from '../state/swap/hooks'
+import { Field } from '../state/swap/types'
 import { useUserPreferredGasPrice } from '../state/user/hooks'
 import { useCurrencyBalance } from '../state/wallet/hooks'
 import { calculateGasMargin } from '../utils'
