+++ src/utils/prices.ts
@@ -34,7 +34,7 @@ import {
   PRICE_IMPACT_MEDIUM,
   PRICE_IMPACT_NON_EXPERT,
 } from '../constants'
-import { Field } from '../state/swap/actions'
+import { Field } from '../state/swap/types'
 
 const Decimal = toFormat(_Decimal)
